+++ src/validators/index.ts
@@ -1,3 +1,4 @@
 export * from './error-message-builder';
+export * from './translations';
 export * from './validation-error';
 export * as Validators from './index-validators';