+++ src/validators/validator-min-max-range-length.ts
@@ -2,7 +2,7 @@ import { isArray, isObject, isString } from 'lodash-es';
 
 import type { FieldBase } from '../field-base';
 
-import { buildErrorMessage } from './error-message-builder';
+import { translatedMessage } from './translations';
 import { RenderContentRef, ValidationErrorRenderContent } from './validation-error';
 import { ValidationFunction, Validator } from './validator';
 
@@ -16,7 +16,7 @@ function toLength(a: any): number {
 
 export class MinLength extends Validator {
   constructor(minLength: number, message?: RenderContentRef) {
-    const msg = message || buildErrorMessage('Length must be larger or equal to **{minLength}**');
+    const msg = message || translatedMessage('MinLength');
     const validationFn: ValidationFunction = (newValue, oldValue, field: FieldBase) => {
       if (toLength(newValue) < minLength) {
         return [
@@ -36,7 +36,7 @@ export class MinLength extends Validator {
 
 export class MaxLength extends Validator {
   constructor(maxLength: number, message?: RenderContentRef) {
-    const msg = message || buildErrorMessage('Length must be less than or equal to **{maxLength}**');
+    const msg = message || translatedMessage('MaxLength');
     const validationFn: ValidationFunction = (newValue, oldValue, field: FieldBase) => {
       if (toLength(newValue) > maxLength) {
         return [
@@ -56,7 +56,7 @@ export class MaxLength extends Validator {
 
 export class LengthInRange extends Validator {
   constructor(minLength: number, maxLength: number, message?: RenderContentRef) {
-    const msg = message || buildErrorMessage('Length must be between **{minLength}** and **{maxLength}**');
+    const msg = message || translatedMessage('LengthInRange');
     const validationFn: ValidationFunction = (newValue, oldValue, field: FieldBase) => {
       const len = toLength(newValue);
       if (len < minLength || len > maxLength) {