+++ src/validators/validator-required.ts
@@ -3,7 +3,7 @@ import { isRef, unref } from 'vue';
 
 import type { FieldBase } from '../field-base';
 
-import { buildErrorMessage } from './error-message-builder';
+import { translatedMessage } from './translations';
 import { RenderContentRef, ValidationErrorRenderContent } from './validation-error';
 import { ValidationFunction, Validator } from './validator';
 
@@ -40,7 +40,7 @@ export default class Required extends Validator {
   constructor(messageOrOptions?: RenderContentRef | RequiredOptions, options?: RequiredOptions) {
     const message = isOptions(messageOrOptions) ? undefined : messageOrOptions;
     const trim = (isOptions(messageOrOptions) ? messageOrOptions : options)?.trim ?? true;
-    const msg = message || buildErrorMessage('Please enter a value');
+    const msg = message || translatedMessage('Required');
     const validationFn: ValidationFunction = (newValue, oldValue, field: FieldBase) => {
       const value = unref(newValue);
       if (toLength(trim && isString(value) ? value.trim() : value) === 0) {